Fix two bugs in dependencies.
1) distinct operator was missing on names.
2) body parameter was being treated as a path parameter.  So the following worked (and was
even covered by a unit test):
POST /stores/{storeId}/order
{
  "storeId": ...
}

A dependency was inferred for the store ID in the body (the same one in the path), because the path
parameter dependency inference was applied to it and the name happened to match

This in some cases is correct, but needs to be modeled explicitly.  For now,
remove this.