Extend spec coverage file with concrete requests and responses.
This change adds an example of a concrete request for which the
data in the spec coverage file is reported.

The goal of this change is that information in the spec coverage file
can be used to troubleshoot failing requests using only:

- the spec coverage file
- backend logs for the service

The diff speccov script has been modified to avoid diffing the
concrete data that is now in the coverage file.