Support header parameters
Changes include:
1. Fuzz header parameters of simple and array types.
2. Support restler_custom_payload_header in dictionary.  This works the
same way as the other types of custom payloads: if a parameter with the
same name is found in the spec, the custom payload from the dictionary
is used instead.
3. Fix bug in the earlier introduced feature that extra custom payload
headers (custom headers) can be added in the dictionary.  This allows the use
case of providing header parameters that are not included in the spec.

Testing:
- added compiler unit tests
- manual testing

Closes #156