Add simple unit test with baseline for smoke test algorithm.
In preparation for merging #233, we need to have a simple unit test with which we can review
the behavior change.  This commit adds such a unit test.

The unit test has 4 requests, A, B, C, D, and E.  A and B do not have any dependencies, while C and D
both depend on both A and B, and E depends on D.
The baseline 'abc_smoke_test_testing_log.txt' captures the current behavior with respect to this grammar.

In the current implementation, all sequences for A, B, C, D will be
rendered "from scratch", but the sequence for E will reuse the 'D' prefix.