Make it possible to analyze flaky sequence failures without parsing full network logs.
This update logs a sample request for sequence failures in the spec coverage file.

The seq.render() code path has been updated to return a meaningful error in case of all failures, and
return the full sequence information in case of sequence failures.  This data can now be added to
the spec coverage log.