Implement injecting custom query parameters.
Use restler_custom_payload_query to inject query params not in the spec.

These parameters will be injected at the end of the query string for every request.

An example use case for restler_custom_payload_query is an application ID or authentication token that needs to be passed in the query.

Closes #287