Fix bugs in example queries.
Bug 1: for requests without bodies, a placeholder needs to be added to correctly represent what is in the grammar, because
the examples are currently tied together via the index only.
Bug 2: incorrect property name

(Note: this means there is no test coverage for having both query and body examples in unit tests.
This needs to be addressed, a separate issue will be opened).