Enable producers from the input payload in the engine
Today, RESTler only supports producer-consumer dependencies by
parsing producers out of the response.

This enables a producer to be specified only as an input parameter.

In the grammar, a new named parameter called 'writer' is added to
the custom_payload_uuid4_suffix parameters.  The generated value
is then written similarly to how values would be parsed out of the response,
to be used in subsequent requests.

(Currently, only uuid4_suffix custom payloads are supported but we may
add support for custom payloads if that is a useful scenario.)

This partially implements #114