Enable input parameters to be used in dependencies (#294)
This change adds support for producers that are not returned in the response.

Currently, this is supported via annotations only (which are specified in the same
way as usual).  For example, if PUT /A/{aId} does not return 'aId' in the
response, prior to this change it was not possible for GET /a/{aId} to use
the same ID if they are uniquely generated.  This is enabled by this change
by adding an annotation:

    {
      "producer_endpoint": "/A/{aId}",
      "producer_method": "PUT",
      "producer_resource_name": "aId",
      "consumer_param": "aId"
    }

In the future, scenarios such as the one above will also be enabled without
having to provide annotations.

See #114