Search for path parameters in the global list of parameters.
It is valid for a specification to declare a required path parameter in the parameters section
without referring to it explicitly.  RESTler was not previously searching in the global parameters list.
After this change, global parameters will be found and assigned even if they are not declared at
the level of the path or method.

Testing: new unit test.