Initial x-ms-paths support. (#297)
With this change, if a specification has an 'x-ms-paths' property in addition to 'paths', the generated
grammar includes the request types from both.  Previously, 'x-ms-paths' was ignored.

The endpoints specified in x-ms-paths which contain a query portion are mapped to a normalized
endpoint that does not contain a query.  RESTler analyzes this normalized endpoint for producer-consumer dependencies as usual, then restores the original endpoint at the time it generates the grammar.

Testing: added unit test corresponding to the in-progress POC.  (Note: this also includes input producers.)