Assign the correct primitive types in dependencies (#303)
The default type for a dependency was being assigned to a string.
However, it should be the original type of the property or parameter, since
this will determine how it is quoted.

(This bug was found when testing examples generated by the engine, directly
from the schema.)

While testing this issue, also fixed how array dependencies are handled, so
now both an array element and the array itself may be a producer.  Added a
unit test for this new behavior, which also confirms that the
dynamic object value is not quoted.