Fix dependency boolean python conversion.
Currently, if a response returns a boolean, it is read as follows:

data = json.loads(response)
boolean_var = str(data["propertyName"])

This causes the json value 'true' to be converted to Python 'True' , which then incorrectly initializes
the dynamic object to 'True' instead of 'true'.

The previous fix for this was incorrect - the correct fix is to only convert the value, if boolean, at
the time it is parsed from the response.