Fix for index out of range error in async utilities. (#352)
The bug was due to missing parenthesis.  Fixed this and another location to use a list instead.

Testing: we have no unit test covering this code path, so this needs to be confirmed with the
original repro.  Automated testing in CI for async should be added, but this will not be
a part of this change.