Fix bug in how query parameters are injected. (#359)
The following case was not working correctly:

Spec: api-version exists
Example 1: api-version exists
Example 2: there is no 'api-version' parameter
The dictionary contains restler_custom_payload_query("api-version")

Expected: each of the above schemas contains just one 'api-version' parameter.  It is injected in Example 2, and replaced in example 1 and the spec schema.
Actual: the api-version parameter was not injected in the third case.