Fix crash in leakage rule checker
Recently, RESTler was changed to return rendering information for sequences with a prefix that failed to re-render (for improved root cause analysis in speccov.json).

However, the checkers should not be executed in this case.  This was missed, and now discovered by triggering a crash in the leakage rule checker.
The fix is to simply not run the checkers in this case (it used to be covered by the check that the rendered sequence is None).

This was not caught because there are only 'smoketest' tests for failures to re-render, not covering the checkers.