Add async wait to the payload body checker.
Today, when valid payloads are sent, async polling is not done and the objects
are leaked.
This fix makes the checker behave the same way as everywhere else - async wait
until the dynamic objects can be parsed out of the response
and garbage collected to avoid quota issues.

Note: this is a quick fix to attempt to resolve the observed quota issues.
It may be desirable to improve this in the future, for example, to avoid
waiting for async polling if all the dynamic objects are available from
names in the request (to avoid waiting to parse the response before deleting it).

The above may not be useful in all cases, since some resources may need to be
fully created before they can be referenced in order to be deleted.