Enable ordering constraints in annotations - compiler changes
This change adds the ability to specify that two requests should be
executed in a specific order.

If this is specified in the global annotations, a dependency variable
will be created in the grammar, and used by the RESTler engine to order
the requests.

Testing: added unit test, manual testing with demo_server.