Enable ordering constraints in the engine
When specified in the 'pre_send' section of the grammar, reader variables
are added to the _consumer list, in the same way as when they are
part of the payload.

Testing: tested manually by adding the following annotations to demo_server,
and confirming that every sequence starts with the GET request.

{"x-restler-global-annotations":
[{
  "producer_endpoint": "/blog/posts",
  "producer_method": "GET",
  "consumer_endpoint": "/blog/posts",
  "consumer_method": "POST"
}]
}