Several bug fixes in the engine schema parser.
Fix #1: The main parsing logic was not supporting restler_custom_payload_header and restler_custom_payload_query
Fix #2: The ad-hoc grammar.py parsing for header parameters that should not be fuzzed
did not handle the case when they are on multiple lines, which is the case for Content-Type after recent changes.
Fix #3: The 'DictionaryCustomPayload' payload type should be included in the list of schema parameters, since these are
the additional parameters injected by the user.

These issues were found testing parameter combinations with demo_server in the presence of the above grammar elements.
Testing: manual testing.

(Lack of automated tests for the schema parser is a known test gap that will be addressed soon.)