Always try to parse responses from GET after async polling (#382)
Previously, if async polling from a PUT response in a certain format was received, that was used as the final
response from which to parse dynamic objects, otherwise the GET request was used.   This did not always work -
in some cases, there were properties missing from the response or an empty body was returned.

After this change, the first response on which the response parser will be invoked will be the one from the GET
request.  If the GET failed or parsing the response of the GET, the response from the PUT will still be tried as before.

Testing: ad-hoc testing using original repro (in progress).