Fix garbage collector to use the specified maximum request execution time.
Using RESTler in CI/CD, it was reported that runs occasionally take 10 minutes while usually finishing in a few seconds.
This was determined to be due to DELETE requests that non-deterministically hung,
and the RESTler GC not using the timeout in the settings, but the default parameter value timeout of 600s.

As part of this fix, the default value of 600 is removed to avoid such issues in the future.