Fix schema cache recursion issue (#425)
The logic was incorrectly caching the leaf instead of root node of the cycle.

Unfortunately, this was not caught in testing because triggering this
in a user-visible way requires a specific ordering of requests (e.g.,
a get or delete followed by a put).  This resulted in response properties
not being extracted from the schema in some cases, causing some missing
producer-consumer dependencies.

Testing: added unit test.