Make 'isRequired' in grammar.json match the spec
Prior to this change, the 'isRequired' flag was always set to 'true' for internal nodes
(arrays and objects with properties).  With the upcoming change to test
combinations of nested parameters, make these match the spec in grammar.json.

Also, document the 'IncludeOptionalParameters' compiler config option.

Added baseline test for grammar.json and grammar.py.