New experimental checker for invalid value fuzzing.
Currently, if a user wants to fuzz invalid values, this is done by adding
values to the main dictionary and testing them through the main algorithm.

This is not the correct approach, because RESTler tests all combinations of
the parameter values in a request, and testing all combinations of the
invalid values is not desirable.  Today, users must re-run RESTler with
custom payloads in order to test large numbers of invalid values
for each parameter efficiently.

This 'invalidvalue' checker is introduced in order to have a separate way to specify the
invalid values.  A static dictionary and/or custom value generators for
invalid values may be provided.
If both a static dictionary and custom value generators are specified,
static values are tested first, then the value generator is used for
the remaining combinations (up to max combinations).

Remaining work:
- Issue warning message is there is an insufficient number of combinations to test
all of the parameters.

Settings file:

- max combinations is max combinations per request.
- One value per parameter is always fuzzed, so if there are 1000
  parameters, this maximum will be exceeded.

    "invalidvalue": {
      "max_combinations": 10,
      "custom_value_generators": "path/to/valuegen.py",
      "custom_dictionary": "path/to/invalid/dict.json"
    }