Fix incorrect handling of empty request IDs.
Empty request ids (endpoint = "/") were not handled correctly.
This was caught by a user observing the error message
"Request from grammar does not exist in the Request Collection!"
The impact of this was that the requests were not fuzzed
with any checkers that use grammar.json (e.g. the payload body checker).

Testing: manual testing with simple repro.