Include all parameters in an example if specified.
Previously, if filtering optional parameters was specified, they were filtered out of the example.
However, examples should be used as-is.  For example, the user may fix a case where the specification
incorrectly marks a parameter as optional by adding an example file (in cases where the specification cannot
be fixed right away).

Testing: added unit test.