Fix payload body checker exception. (#508)
Closes #505.
The payload body checker filters out the dictionary values for restler_fuzzable_string when calculating coverage metrics based on (error code, error message). The issue above is that when a dynamic generator is specified instead of a static list, this filter can no longer be initialized once at the start of fuzzing. Moreover, the filter does not address every type of property (e.g. custom payloads), so it is incomplete.

As a workaround, simply skip the dynamic generator when initializing the static list of strings to exclude from analysis as above.

The root cause fix should be addressed by either implementing a better bucketization scheme (similar to results analyzer), or by having the analysis exclude all fuzzables by traversing the initialized python grammar elements (similar to invalid value checker).