Cover additional case with x-ms-path query parameters. (#519)
In some specifications, the query parameter names that are part of the path
are declared.  For example, for the endpoint /customer&operation=create, the
'operation' is declared as a query parameter.  This was not previously handled -
RESTler now filters these out of the query parameters it appends after the path.