Re-implement demo_server using FastAPI
This migration is being done in an attempt to
improve recurring flask-restplus related dependency breaks
and to resolve a recent vulnerability alert in a dependent component.

Note: coverage collection that was done on the flask-based
demo_server was not re-tested on the FastAPI-based one, and hence may not work.
This was only used for initial fuzzing experiments
(from which the graphs were generated for the first research paper),
hence it should not impact anyone actively using demo_server.

Testing: end to end tests that use demo_server passing in CI pipeline.

TODO:
- update the TutorialDemoServer documentation
- update README