Background: the grammar.json schema parser and python grammar generator were initially developed for the payload body checker. Over time, other parts of the RESTler engine began to use this parser and python request definition generator code, but it was never thoroughly tested.
This change makes several fixes and enhancements to the schema parser and request definition generator as well as adds a sanity test for a large portion of the existing functionality.

Specific changes include:

added 'get_original_blocks' function to each param object, which is supposed to generate
exactly the python grammar corresponding to the original json grammar.
The existing 'get_blocks' and 'get_fuzzing_blocks' do not satisfy this requirement.
They are kept for backwards compatibility for the payload body checker.
fixed several bugs in schema parsing
added missing elements to the data types
added the ability to filter schema nodes using a filter function
(this will be used in a subsequent change)
Testing:

new unit test
manual real-world testing of Test, test_all_combinations, FuzzLean on several services