Fix missing producer-consumer dependency writer
The RESTler compiler did not handle the case when
the producer of a user-specified producer-consumer annotation was also a
consumer that was dynamically inferred by RESTler.

This update handles this case by propagating the original inferred producer
to the annotation consumer.  An ordering constraint must also be added.

Closes #562

Testing:
- added unit test for the repro and another related case where
the user-specified producer is also an inferred uuid suffix.