Add new option to restler driver to generate a job configuration
The purpose of this new command line option is to allow the user to get
an initial set of all the files that may need to be customized  to improve RESTler coverage.

Usage:
PS > .\restler\restler.exe --generate_config <list of file or directory paths> --output_dir <directory to create for the job configuration>

If the output_dir is not specified, a sub-directory in the cwd will be created.

Testing:
- manual testing, test coverage will be added in a follow-up change