Add new settings to include and exclude specific requests.
This is required for maintaining specific endpoints that should not be exercised, and is a cleaner way
to do so than the existing path_regex switch.

Also:
- fixed a bug in path_regex.
- implemented the TODO to always order the requests (since the ordering changed and some test baselines
had to be updated, just went ahead and updated them all.

Testing:
- manual testing on demo_server.
- updated unit tests

Partially implements #282.  A new property will be added separately to completely exclude DELETEs from running.