Fix writer variable update bug
'render_iter' should only generate combinations, and in particular should not have any side effects that trigger garbage collection.
This was missed in the earlier implementation, so GC would get triggered even if the request was never executed, because
dynamic objects were added to the cache in render().

Instead, writer variables should only be updated after the request has succeeded (returned a valid code), which means that
the dynamic object exists.