Fix writer variable update bug (#663)
'render_iter' should only generate combinations - it should not have any side effects. Garbage collection was being triggered in the presence of writer variables even if the request was never executed, because dynamic objects were added to the cache in render(). This was missed in the original implementation of writer variables.

Instead, writer variables should only be updated after the request has succeeded (returned a valid code), indicating that the dynamic object exists.

Testing: existing unit tests and manual inspection of GC logs.