New report to speed up improving coverage
Today, it can be slow and painful to figure out some basic questions during coverage investigations by reading the RESTler output files, without writing additional scripts.  For example:

1) Which request, if fixed, would unblock the highest number of requests?  In other words, which request should I work on fixing first?

2)What are all the failed requests that were actually tried (not skipped because of a dependency)?

3) What is the sequence of concrete requests that were sent for each failure?
Today, this needs to be gathered by navigating and copy-pasting in speccov.json, or by finding the same sequence in the network logs (which is not easy to correlate with the speccov file, since request/sequence IDs are not printed there).

This change includes the following improvements to address these issues:

1. The engine writes a second file in addition to speccov.json, speccov-min.json, which includes the raw request and response.
This leaves the speccov.json file unmodified, since speccov.json format is optimized for directly uploading to a database for further
analysis, whereas speccov-min is generated to provide a quick way to generate a report that is more user-friendly to inspect directly.

2. The engine now prints all combinations to the speccov.json file (up to a maximum), then post-processes this file to determine the data for each request, and provides the first request as the sample request.  This allows failures for user-specified examples to be quickly investigated - the last request, on the contrary, is usually a RESTler-generated request (e.g. with 'fuzzstring' plugged in for all string parameters).

Closes #637

3. This new 'speccov-min.json' is post-processed by the RESTler driver to produce a new text file, which contains the failed requests, prioritized by the number of dependent requests.  For each failed request, RESTler includes the full sequence of requests and responses up to the failure.

4. Fixed several bugs and missing information in the spec coverage file.