Fix date serialization bug in the mutations dictionary.
In some cases, dates were being reformatted during compilation.  For example, if a custom payload specified a date as follows -

"2023-03-13T10:36:56.0000000Z"

It was reformatted to be:

"03/13/2023 10:36:56"

This was due to using the default option for Json deserialization - the fix is to specify DateParseHandling.None
in the serializer settings.

Since this issue could occur in many places (examples, grammar, mutations dictionary), this change updates the serialization
throughout the code to use the common module with this option.

Testing:
- added unit test

TODO: add test with all the elements of dict.