Fix telemetry regression (#747)
This severe regression was introduced when fixing issue #270 (commit 5a7d240).
The telemetry client sends telemetry in its Dispose(), which the above commit broke
because it modified main to terminate with `exit` instead of normal termination.

Since there may have been a reason to use `exit` at the time, this fix modifies the
telemetry client to flush after sending top-level started and finished events.  This is desirable
anyway, so the RestlerStarted events are sent even if RESTler does not finish (e.g. machine restart).

Testing: manual testing for all telemetry events to be present in AppInsights