Enable setting equality constraints on uuid4 suffix payloads
Some request types have payloads that has a unique value generated via uuid4_suffix (e.g. a PUT request that
must create a new ID every time), and require the same value to be passed in multiple places in the payload.
RESTler did not previously have a way to specify the equality constraints on such payload parameters or properties.

This change enables this via an annotation, for example:

```
    {
      "producer_endpoint": "/archive/{archiveId}",
      "producer_method": "PUT",
      "consumer_endpoint": "/archive/{archiveId}",
      "consumer_method": "PUT",
      "producer_resource_name": "archiveId",
      "consumer_param": "/id"
    }
```

Testing:
- added new unit test

Partially implements #89