Fix missing producer_timing_delay.
This bug was found when specifying a `producer_timing_delay` for a resource that was
also specified as `create_once`.  Here, the length of the sequence was 1, and RESTler
proceeded fuzzing with the `create_once` resource in a not-ready state.

By default, change the behavior to always wait after the resource creation.