Support replay mode with garbage collection
This change enables a new 'replay' mode option that replays requests from the trace database.

Changes include:

1. Refactor driver to consume sequences from different sources (with trace database another supported source of specific
sequences, similar to the existing smoke test mode)

2. Generate 'replay blocks' as part of payload rendering.  Replay blocks contain dynamic objects, but do not support
custom payloads (this is future work).

3. Enables replay to be based on replay blocks

With this change, RESTler generates sequences from replay blocks if available,
or from raw request/responses (similar to the existing replay mode) when they are not available.

The current version of replay is based on the rendered data sent at the time of
original sequence rendering.  This has several limitations, such as not being able to
garbage-collect resources or plugging in custom payloads.

In particular, when a bug is reproduced in RESTler today, the same replay mechanism is used, which
means that GC does not collect resources created while reproducing the bug.  Moreover, re-running
the same sequence without GC will not work if the resource has not been deleted if a resource-generating
request has a unique ID parameter (e.g. PUT /resource/{resourceId}), causing false negative non-reproducible
bugs.

This change addresses this issue by implementing replay blocks-based replay, which is invoked when reproducing bugs.

Note: the existing replay based on .replay.txt files will also be able to use this new mechanism if a grammar is provided,
but this is not implemented as part of this change.

4. Added a setting to filter the origin during DB replay.

To only replay specific origins, add the following to the settings file:
  "replay": {
    "include_origins": ["main_driver", "InvalidValueChecker"]
  }

Testing:
- manual testing of demo_server replay as follows:

1) Run 'test' task and generate trace database
>restler.exe test --grammar_file .\Compile\grammar.py --dictionary_file .\Compile\dict.json --host localhost --target_port 8888  --settings .\compile\engine_settings.json --no_ssl

Engine settings:
{
  "use_trace_database": true,
  "trace_database": {
    "root_dir": "d:\\demo_server\\trace_databases",
  },
}

2) Run 'replay' task from above trace database

>restler.exe replay --replay_log ./trace_data.ndjson --grammar_file .\Compile\grammar.py --dictionary_file .\Compile\dict.json --host localhost --target_port 8888  --settings .\compile\engine_settings.json --no_ssl

- updated unit tests