Enable traceDB-based replay without the request text
For use cases when a trace database is saved and re-played,
this change allows using only replay blocks without the rendered requests.
This is being added for scenarios where only replay blocks are saved during
custom serialization.

The rendered requests continue to be logged by default, since they are
useful for debugging.

Testing:
- manual testing

- new unit test that replays from ndjson with vs. without the request text