Fix Content-Type custom payload bugs
When a custom payload body is used to work around unsupported content types, the Content-Type
header must also be modified via the dictionary.  This functionality had several bugs, which
are fixed in this change.  After this update, using either `restler_custom_payload_header` or
`restler_custom_payload` with the request-specific syntax for dictionary payloads now works
as expected, for example:

  "restler_custom_payload": {
    "/stores/{storeId}/order/post/Content-Type": [ "xml" ]
  }

This change also fixes the same bugs with per-request syntax with restler_custom_payload_query.

Testing:
- manual testing with demo_server
- added new compiler tests