Update default fuzzing timeout to 7 days
Currently, the fuzzing timeout is 1 hour by default.  This causes incorrect/misleading results
for more complex services that require a longer fuzzing time,
if timeout is accidentally omitted due to a configuration issue.

Providing a very long timeout in an attempt to mitigate this issue. (Note: this was already the default
timeout in the RESTler engine - this change is only for the driver).

To restore the default timeout, add the following to the engine settings:

'time_budget': 1