Search mode default and doc updates
Per discussion in #823, updating the README and updating the default mode to be `bfs-cheap`.

Originally, the reasoning behind the default value of `bfs-fast` was that it provides different coverage from `fuzz-lean`
(which is similar to `bfs-cheap` in that it does not exercise all of the renderings).
However, many users only want to run one fuzzing mode (e.g. for evaluation purposes, having just one job in CI/CD), and
that is typically the `fuzz` mode (since we document it to provide the best coverage).

Updating the documentation to reflect this.