Fix Windows only new infrastructure failure (#958)
This started occurring due to a compilation error in the F# code base, which was due to a bug in an unused
data structure.  This is only caught by .NET 9 SDK, which was being used on Windows but not on Linux.

Changes include:
- fix the bug
- add error handling and debugging to the github action.  The error handling
was automatically done in ADO pipelines, and adding it was missed during the migration to
github actions.
- pin the .NET SDK to the version that is currently supported.